update to jquery.jtable.editinline.js
Changed the ajax call success to inform the user if editinline gets an error message back from the server while inserting data. instead of just hanging there and looking like the update worked.

Used along with 

recordUpdated: function (event, data) { $('.jtable-child-table-container').jtable('reload'); }

The user is notified but the row also reflects it.

Maybe there is a way to just revert the cell back to the old content here but I am too lazy to look into it.